requirements.txt
# Defect Dojo - Official Requirements
# Based on https://github.com/DefectDojo/django-DefectDojo/blob/dev/requirements.txt

# Core Django
Django>=3.2,<4.2
djangorestframework>=3.12.4,<3.15
django-cors-headers>=3.7.0,<4.0
django-filter>=2.4.0,<23.0
django-environ>=0.4.5,<0.11
django-axes>=5.0.12,<6.0
django-crispy-forms>=1.11.2,<2.0
crispy-bootstrap5>=0.6,<0.7
django-extensions>=3.1.3,<3.2
django-import-export>=2.5.0,<2.6
django-multiselectfield>=0.1.12,<0.2
django-notifications-hq>=1.6.0,<1.7
django-simple-history>=2.12.0,<3.0
django-taggit>=1.3.0,<2.0
django-tastypie>=0.14.2,<0.15
django-tastypie-swagger>=0.1.4,<0.2
django-watson>=1.5.5,<1.6
django-celery-beat>=2.2.1,<2.3
django-celery-results>=2.2.0,<2.3

# Database
psycopg2-binary>=2.8.6,<2.9
mysqlclient>=2.0.3,<2.1

# Web server
gunicorn>=20.1.0,<20.2
whitenoise>=5.3.0,<6.0

# Background tasks
celery>=5.2.3,<5.3
redis>=3.5.3,<3.6

# Security
cryptography>=3.4.7,<3.5
PyJWT>=2.1.0,<2.2

# API and integrations
requests>=2.25.1,<2.26
lxml>=4.6.3,<4.7
openpyxl>=3.0.7,<3.1
xlsxwriter>=1.4.3,<1.5
python-docx>=0.8.11,<0.9
python-pptx>=0.6.18,<0.7
reportlab>=3.5.68,<3.6
Pillow>=8.2.0,<8.3

# Monitoring and logging
sentry-sdk>=1.1.0,<1.2
structlog>=21.1.0,<21.2

# Utilities
python-dateutil>=2.8.1,<2.9
python-magic>=0.4.24,<0.5
python-magic-bin>=0.4.14; sys_platform == 'win32'
markdown>=3.3.4,<3.4
bleach>=3.3.0,<3.4
html5lib>=1.1,<1.2

# Email
django-anymail[mailgun]>=8.2,<8.3

# File handling
django-storages[boto3]>=1.11.1,<1.12
boto3>=1.17.49,<1.18

# Testing and development
factory-boy>=3.2.0,<3.3
faker>=8.1.0,<8.2
pytest>=6.2.4,<6.3
pytest-django>=4.4.0,<4.5
pytest-cov>=2.12.1,<2.13

# Security scanning tools (for DevSecOps)
bandit>=1.7.0,<1.8
safety>=1.10.3,<1.11
semgrep>=1.0.0,<1.1 